Extract BatchedTrigEncodeOptimizer and add secondary encode cache
The batch/fused encoding algorithm was inline in RichChkTrigTranscoder,
making the transcoder very hard to read. This isolates all encoding
complexity into BatchedTrigEncodeOptimizer so the transcoder only holds
decode logic and a two-line encode() delegation.

Also adds a secondary encode cache in RichChkIo keyed on the trigger
structural fingerprint + amounts hash, allowing repeated encode_chk calls
with structurally-identical (but distinct object identity) RichTrigSections
to return cached results without re-encoding.
