+++ src/richchk/transcoder/richchk/transcoders/richchk_trig_transcoder.py
@@ -1,7 +1,4 @@
 """Decode the TRIG - Triggers section."""
-import array
-import struct
-import sys
 from typing import Any, ClassVar, Optional, Union, cast
 
 from ....model.chk.trig.decoded_player_execution import DecodedPlayerExecution
@@ -14,9 +11,6 @@ from ....model.richchk.richchk_encode_context import RichChkEncodeContext
 from ....model.richchk.trig.actions.flags.trigger_action_flags import (
     _DEFAULT_TRIGGER_ACTION_FLAGS,
 )
-from ....model.richchk.trig.actions.preserve_trigger_action import PreserveTrigger
-from ....model.richchk.trig.actions.set_deaths_action import SetDeathsAction
-from ....model.richchk.trig.conditions.deaths_condition import DeathsCondition
 from ....model.richchk.trig.conditions.flags.trigger_condition_flags import (
     _DEFAULT_TRIGGER_CONDITION_FLAGS,
 )
@@ -36,14 +30,13 @@ from ....transcoder.richchk.richchk_section_transcoder_factory import (
 )
 from ....util import logger
 from .helpers.richchk_enum_transcoder import RichChkEnumTranscoder
+from .trig.batched_trig_encode_optimizer import BatchedTrigEncodeOptimizer
 from .trig.rich_trigger_action_transcoder_factory import (
     RichTriggerActionTranscoderFactory,
 )
 from .trig.rich_trigger_condition_transcoder_factory import (
     RichTriggerConditionTranscoderFactory,
 )
-
-_IS_BIG_ENDIAN: bool = sys.byteorder == "big"
 
 
 class RichChkTrigTranscoder(
@@ -52,52 +45,9 @@ class RichChkTrigTranscoder(
     chk_section_name=DecodedTrigSection.section_name(),
 ):
 
-    _NUM_CONDITIONS_PER_TRIGGER = 16
-    _NUM_ACTIONS_PER_TRIGGER = 64
-
-    # Binary format constants (mirrored from ChkTrigTranscoder)
-    _CONDITION_FORMAT = "3I H 4B H"
-    _ACTION_FORMAT = "6I H 3B B H"
-    _PLAYER_EXECUTION_FORMAT = "I 27B B"
-    _NUM_BYTES_PER_CONDITION = 20
-    _NUM_BYTES_PER_ACTION = 32
-    _NUM_BYTES_PER_PE = 32
-    _CONDS_SECTION_SIZE = _NUM_BYTES_PER_CONDITION * _NUM_CONDITIONS_PER_TRIGGER  # 320
-    _ACTS_SECTION_SIZE = _NUM_BYTES_PER_ACTION * _NUM_ACTIONS_PER_TRIGGER  # 2048
-    _NUM_BYTES_PER_TRIGGER = (
-        _CONDS_SECTION_SIZE + _ACTS_SECTION_SIZE + _NUM_BYTES_PER_PE
-    )
-
-    # Inline IDs for the most common condition/action types
-    _DEATHS_CID: int = DeathsCondition.condition_id().id
-    _SET_DEATHS_AID: int = SetDeathsAction.action_id().id
-    _PRESERVE_AID: int = PreserveTrigger.action_id().id
-
-    # Pre-packed 32 bytes for a PreserveTrigger action (all zeros except action_id byte)
-    _PRESERVE_TRIGGER_BYTES: ClassVar[bytes] = struct.pack(
-        "6I H 3B B H", 0, 0, 0, 0, 0, 0, 0, PreserveTrigger.action_id().id, 0, 0, 0, 0
-    )
-    _COND_STRUCT: ClassVar[struct.Struct] = struct.Struct(_CONDITION_FORMAT)
-    _ACT_STRUCT: ClassVar[struct.Struct] = struct.Struct(_ACTION_FORMAT)
-    _PE_STRUCT: ClassVar[struct.Struct] = struct.Struct(_PLAYER_EXECUTION_FORMAT)
-
     _action_type_cache: ClassVar[dict[Any, Any]] = {}
     _condition_type_cache: ClassVar[dict[Any, Any]] = {}
-    _fused_pe_bytes_cache: ClassVar[dict[Any, Any]] = {}
-    _I_STRUCT: ClassVar[struct.Struct] = struct.Struct("<I")
-    _template_cache: ClassVar[
-        dict[Any, Any]
-    ] = (
-        {}
-    )  # id(triggers) → (nz_pairs, template_bytes, cond_patches, act_patches) or False
-    _template_fingerprint_cache: ClassVar[
-        dict[Any, Any]
-    ] = (
-        {}
-    )  # structural fingerprint → (nz_pairs, template_bytes, cond_patches, act_patches)
-    _nz_bufs_cache: ClassVar[
-        dict[Any, Any]
-    ] = {}  # (id(nz_pairs), n) → [(pos, buf)] pre-built strided write buffers
+    _optimizer: ClassVar[BatchedTrigEncodeOptimizer] = BatchedTrigEncodeOptimizer()
 
     _EMPTY_ACTION: DecodedTriggerAction = DecodedTriggerAction(
         _location_id=0,
@@ -269,493 +219,22 @@ class RichChkTrigTranscoder(
         rich_chk_section: RichTrigSection,
         rich_chk_encode_context: RichChkEncodeContext,
     ) -> DecodedTrigSection:
-        raw_bytes = self._fused_encode_to_bytes(
-            rich_chk_section, rich_chk_encode_context
-        )
+        raw_bytes = self._optimizer.encode(rich_chk_section, rich_chk_encode_context)
         return DecodedTrigSection(_triggers=[], _raw_data=raw_bytes)
 
-    @staticmethod
-    def _compute_trigger_fingerprint(
-        conds0: Any, acts0: Any, players0: Any
-    ) -> tuple[Any, ...]:
-        return (
-            tuple((c._group, c._unit, c._comparator) for c in conds0),
-            tuple(
-                (a._group, a._unit, a._amount_modifier)
-                if type(a) is SetDeathsAction
-                else ()
-                for a in acts0
-            ),
-            players0,
-        )
-
-    def _build_template_info(self, triggers: Any) -> Any:
-        """Try to build a per-trigger template for uniform
-        DeathsCondition+SetDeaths+Preserve batches.
-
-        Returns (template_bytes, cond_patches, act_patches) if all triggers share the
-        same non-amount fields, else False. cond_patches: tuple of (cond_index,
-        byte_offset_within_trigger) for varying amounts act_patches: tuple of
-        (act_index, byte_offset_within_trigger) for varying amounts
-        """
-        if not triggers:
-            return False
-        t0 = triggers[0]
-        conds0 = t0._conditions
-        acts0 = t0._actions
-        players0 = t0._players
-        n_conds = len(conds0)
-        n_acts = len(acts0)
-
-        for c in conds0:
-            if (
-                type(c) is not DeathsCondition
-                or c._flags is not _DEFAULT_TRIGGER_CONDITION_FLAGS
-            ):
-                return False
-        for a in acts0:
-            at = type(a)
-            if at is SetDeathsAction:
-                if a._flags is not _DEFAULT_TRIGGER_ACTION_FLAGS:
-                    return False
-            elif at is not PreserveTrigger:
-                return False
-
-        fingerprint = self._compute_trigger_fingerprint(conds0, acts0, players0)
-        fp_result = self._template_fingerprint_cache.get(fingerprint)
-        if fp_result is not None:
-            return fp_result
-
-        cond_amounts_vary = [False] * n_conds
-        act_amounts_vary = [False] * n_acts
-        check_count = min(len(triggers), 200)
-        for i_check in range(check_count):
-            t = triggers[i_check]
-            if len(t._conditions) != n_conds or len(t._actions) != n_acts:
-                return False
-            if t._players != players0:
-                return False
-            for i in range(n_conds):
-                c = t._conditions[i]
-                c0 = conds0[i]
-                if (
-                    type(c) is not DeathsCondition
-                    or c._flags is not _DEFAULT_TRIGGER_CONDITION_FLAGS
-                    or c._group is not c0._group
-                    or c._unit is not c0._unit
-                    or c._comparator is not c0._comparator
-                ):
-                    return False
-                if c._amount != c0._amount:
-                    cond_amounts_vary[i] = True
-            for j in range(n_acts):
-                a = t._actions[j]
-                a0 = acts0[j]
-                at = type(a)
-                if at is not type(a0):
-                    return False
-                if at is SetDeathsAction:
-                    if (
-                        a._flags is not _DEFAULT_TRIGGER_ACTION_FLAGS
-                        or a._group is not a0._group
-                        or a._unit is not a0._unit
-                        or a._amount_modifier is not a0._amount_modifier
-                    ):
-                        return False
-                    if a._amount != a0._amount:
-                        act_amounts_vary[j] = True
-                elif at is not PreserveTrigger:
-                    return False
-
-        trig_sz = self._NUM_BYTES_PER_TRIGGER
-        template = bytearray(trig_sz)
-        pack_into_cond = self._COND_STRUCT.pack_into
-        pack_into_act = self._ACT_STRUCT.pack_into
-        pack_into_pe = self._PE_STRUCT.pack_into
-        conds_sz = self._CONDS_SECTION_SIZE
-        deaths_cid = self._DEATHS_CID
-        set_deaths_aid = self._SET_DEATHS_AID
-
-        cond_patches = []
-        cond_off = 0
-        for i, c in enumerate(conds0):
-            tmpl_amount = 0 if cond_amounts_vary[i] else c._amount
-            pack_into_cond(
-                template,
-                cond_off,
-                0,
-                c._group._id,
-                tmpl_amount,
-                c._unit._id,
-                c._comparator._id,
-                deaths_cid,
-                0,
-                0,
-                0,
-            )
-            if cond_amounts_vary[i]:
-                cond_patches.append((i, cond_off + 8))
-            cond_off += self._NUM_BYTES_PER_CONDITION
-
-        act_patches = []
-        act_off = conds_sz
-        for j, a in enumerate(acts0):
-            at = type(a)
-            if at is SetDeathsAction:
-                tmpl_amount = 0 if act_amounts_vary[j] else a._amount
-                pack_into_act(
-                    template,
-                    act_off,
-                    0,
-                    0,
-                    0,
-                    0,
-                    a._group._id,
-                    tmpl_amount,
-                    a._unit._id,
-                    set_deaths_aid,
-                    a._amount_modifier._id,
-                    0,
-                    0,
-                    0,
-                )
-                if act_amounts_vary[j]:
-                    act_patches.append((j, act_off + 20))
-            elif at is PreserveTrigger:
-                template[act_off : act_off + 32] = self._PRESERVE_TRIGGER_BYTES
-            act_off += self._NUM_BYTES_PER_ACTION
-
-        pe_sz = self._NUM_BYTES_PER_PE
-        pe_base = conds_sz + self._ACTS_SECTION_SIZE
-        pe_cache = self._fused_pe_bytes_cache
-        pe_bytes = pe_cache.get(players0)
-        if pe_bytes is None:
-            pe_execution_cache = self._player_execution_cache
-            pe = pe_execution_cache.get(players0)
-            if pe is None:
-                player_flags = [0] * 27
-                for player_id in players0:
-                    player_flags[player_id._id] = 1
-                pe = DecodedPlayerExecution(
-                    _execution_flags=0,
-                    _player_flags=player_flags,
-                    _current_action_index=0,
-                )
-                pe_execution_cache[players0] = pe
-            buf = bytearray(pe_sz)
-            pack_into_pe(
-                buf, 0, pe._execution_flags, *pe._player_flags, pe._current_action_index
-            )
-            pe_bytes = bytes(buf)
-            pe_cache[players0] = pe_bytes
-        template[pe_base : pe_base + pe_sz] = pe_bytes
-
-        fp_nz_pairs = tuple((i, b) for i, b in enumerate(template) if b)
-        result = (fp_nz_pairs, bytes(template), tuple(cond_patches), tuple(act_patches))
-        self._template_fingerprint_cache[fingerprint] = result
-        return result
-
-    def _fused_encode_to_bytes(
-        self,
-        rich_chk_section: RichTrigSection,
-        context: RichChkEncodeContext,
-    ) -> bytes:
-        triggers = rich_chk_section.triggers
-        trig_sz = self._NUM_BYTES_PER_TRIGGER
-        total = len(triggers) * trig_sz
-
-        if triggers:
-            trig_id = id(triggers)
-            template_info = self._template_cache.get(trig_id)
-            if template_info is None:
-                template_info = self._build_template_info(triggers)
-                self._template_cache[trig_id] = template_info
-            if template_info is not False:
-                nz_pairs, template_bytes, cond_patches, act_patches = template_info
-                n = len(triggers)
-                # Zero-fill + strided writes: faster than template * n for sparse templates
-                # because bytearray(N) uses calloc (near-zero cost for reused pages) while
-                # template * n does a full 12 MB copy of mostly-zero content.
-                if nz_pairs and len(nz_pairs) <= 64:
-                    nz_key = (id(nz_pairs), n)
-                    nz_bufs = self._nz_bufs_cache.get(nz_key)
-                    if nz_bufs is None:
-                        nz_bufs = [(pos, bytes([val]) * n) for pos, val in nz_pairs]
-                        self._nz_bufs_cache[nz_key] = nz_bufs
-                    data = bytearray(n * trig_sz)
-                    for pos, buf in nz_bufs:
-                        data[pos::trig_sz] = buf
-                else:
-                    data = bytearray(template_bytes * n)
-                if cond_patches or act_patches:
-                    if len(cond_patches) == 1 and not act_patches:
-                        cond_idx, cond_off_rel = cond_patches[0]
-                        amounts = [t._conditions[cond_idx]._amount for t in triggers]
-                        amt = array.array("I", amounts)
-                        if _IS_BIG_ENDIAN:
-                            amt.byteswap()
-                        ab = amt.tobytes()
-                        data[cond_off_rel::trig_sz] = ab[0::4]
-                        data[cond_off_rel + 1 :: trig_sz] = ab[1::4]
-                        data[cond_off_rel + 2 :: trig_sz] = ab[2::4]
-                        data[cond_off_rel + 3 :: trig_sz] = ab[3::4]
-                    else:
-                        for cond_idx, cond_off in cond_patches:
-                            amounts = [
-                                t._conditions[cond_idx]._amount for t in triggers
-                            ]
-                            amt = array.array("I", amounts)
-                            if _IS_BIG_ENDIAN:
-                                amt.byteswap()
-                            ab = amt.tobytes()
-                            data[cond_off::trig_sz] = ab[0::4]
-                            data[cond_off + 1 :: trig_sz] = ab[1::4]
-                            data[cond_off + 2 :: trig_sz] = ab[2::4]
-                            data[cond_off + 3 :: trig_sz] = ab[3::4]
-                        for act_idx, act_off in act_patches:
-                            amounts = [t._actions[act_idx]._amount for t in triggers]
-                            amt = array.array("I", amounts)
-                            if _IS_BIG_ENDIAN:
-                                amt.byteswap()
-                            ab = amt.tobytes()
-                            data[act_off::trig_sz] = ab[0::4]
-                            data[act_off + 1 :: trig_sz] = ab[1::4]
-                            data[act_off + 2 :: trig_sz] = ab[2::4]
-                            data[act_off + 3 :: trig_sz] = ab[3::4]
-                return data
-
-        data = bytearray(total)
-        pack_into_cond = self._COND_STRUCT.pack_into
-        pack_into_act = self._ACT_STRUCT.pack_into
-        pack_into_pe = self._PE_STRUCT.pack_into
-        conds_sz = self._CONDS_SECTION_SIZE
-        acts_sz = self._ACTS_SECTION_SIZE
-        num_bytes_per_cond = self._NUM_BYTES_PER_CONDITION
-        num_bytes_per_act = self._NUM_BYTES_PER_ACTION
-        pe_sz = self._NUM_BYTES_PER_PE
-        deaths_cid = self._DEATHS_CID
-        set_deaths_aid = self._SET_DEATHS_AID
-        preserve_bytes = self._PRESERVE_TRIGGER_BYTES
-        preserve_type = PreserveTrigger
-        deaths_type = DeathsCondition
-        set_deaths_type = SetDeathsAction
-        decoded_cond_type = DecodedTriggerCondition
-        decoded_act_type = DecodedTriggerAction
-        act_cache = self._action_type_cache
-        cond_cache = self._condition_type_cache
-        pe_cache = self._fused_pe_bytes_cache
-        pe_execution_cache = self._player_execution_cache
-
-        offset = 0
-        for trigger in triggers:
-            # --- conditions ---
-            cond_off = offset
-            for condition in trigger._conditions:
-                c_type = type(condition)
-                if c_type is deaths_type:
-                    dc_cond = cast(DeathsCondition, condition)
-                    f = dc_cond._flags
-                    if f is _DEFAULT_TRIGGER_CONDITION_FLAGS:
-                        flags_byte = 0
-                    else:
-                        flags_byte = (
-                            int(f.unknown)
-                            | (int(f.disabled) << 1)
-                            | (int(f.always_display) << 2)
-                            | (int(f.unit_properties_is_used) << 3)
-                            | (int(f.unit_type_is_used) << 4)
-                        )
-                    pack_into_cond(
-                        data,
-                        cond_off,
-                        0,
-                        dc_cond._group._id,
-                        dc_cond._amount,
-                        dc_cond._unit._id,
-                        dc_cond._comparator._id,
-                        deaths_cid,
-                        0,
-                        flags_byte,
-                        0,
-                    )
-                elif c_type is decoded_cond_type:
-                    raw_cond = cast(DecodedTriggerCondition, condition)
-                    pack_into_cond(
-                        data,
-                        cond_off,
-                        raw_cond._location_id,
-                        raw_cond._group,
-                        raw_cond._quantity,
-                        raw_cond._unit_id,
-                        raw_cond._numeric_comparison_operation,
-                        raw_cond._condition_id,
-                        raw_cond._numeric_comparand_type,
-                        raw_cond._flags,
-                        raw_cond._mask_flag,
-                    )
-                else:
-                    rich_cond = cast(RichTriggerCondition, condition)
-                    transcoder = cond_cache.get(c_type)
-                    if transcoder is None:
-                        cid = rich_cond.condition_id()
-                        if RichTriggerConditionTranscoderFactory.supports_transcoding_condition(
-                            cid
-                        ):
-                            _f = RichTriggerConditionTranscoderFactory
-                            transcoder = _f.make_rich_trigger_condition_transcoder(cid)
-                            cond_cache[c_type] = transcoder
-                        else:
-                            raise ValueError(
-                                f"No transcoder for condition type: {c_type}"
-                            )
-                    if rich_cond.flags is _DEFAULT_TRIGGER_CONDITION_FLAGS:
-                        dc = transcoder._encode(rich_cond, context)
-                    else:
-                        dc = transcoder.encode(rich_cond, context)
-                    pack_into_cond(
-                        data,
-                        cond_off,
-                        dc._location_id,
-                        dc._group,
-                        dc._quantity,
-                        dc._unit_id,
-                        dc._numeric_comparison_operation,
-                        dc._condition_id,
-                        dc._numeric_comparand_type,
-                        dc._flags,
-                        dc._mask_flag,
-                    )
-                cond_off += num_bytes_per_cond
-
-            # --- actions ---
-            act_base = offset + conds_sz
-            act_off = act_base
-            for action in trigger._actions:
-                a_type = type(action)
-                if a_type is set_deaths_type:
-                    sd_act = cast(SetDeathsAction, action)
-                    fa = sd_act._flags
-                    if fa is _DEFAULT_TRIGGER_ACTION_FLAGS:
-                        flags_byte = 0
-                    else:
-                        flags_byte = (
-                            int(fa.ignore_wait_or_transmission_once)
-                            | (int(fa.disabled) << 1)
-                            | (int(fa.always_display) << 2)
-                            | (int(fa.unit_properties_is_used) << 3)
-                            | (int(fa.unit_type_is_used) << 4)
-                        )
-                    pack_into_act(
-                        data,
-                        act_off,
-                        0,
-                        0,
-                        0,
-                        0,
-                        sd_act._group._id,
-                        sd_act._amount,
-                        sd_act._unit._id,
-                        set_deaths_aid,
-                        sd_act._amount_modifier._id,
-                        flags_byte,
-                        0,
-                        0,
-                    )
-                elif a_type is preserve_type:
-                    data[act_off : act_off + 32] = preserve_bytes
-                elif a_type is decoded_act_type:
-                    raw_act = cast(DecodedTriggerAction, action)
-                    pack_into_act(
-                        data,
-                        act_off,
-                        raw_act._location_id,
-                        raw_act._text_string_id,
-                        raw_act._wav_string_id,
-                        raw_act._time,
-                        raw_act._first_group,
-                        raw_act._second_group,
-                        raw_act._action_argument_type,
-                        raw_act._action_id,
-                        raw_act._quantifier_or_switch_or_order,
-                        raw_act._flags,
-                        raw_act._padding,
-                        raw_act._mask_flag,
-                    )
-                else:
-                    rich_act = cast(RichTriggerAction, action)
-                    transcoder = act_cache.get(a_type)
-                    if transcoder is None:
-                        aid = rich_act.action_id()
-                        if RichTriggerActionTranscoderFactory.supports_transcoding_trig_action(
-                            aid
-                        ):
-                            _af = RichTriggerActionTranscoderFactory
-                            transcoder = _af.make_rich_trigger_action_transcoder(aid)
-                            act_cache[a_type] = transcoder
-                        else:
-                            raise ValueError(f"No transcoder for action type: {a_type}")
-                    if rich_act.flags is _DEFAULT_TRIGGER_ACTION_FLAGS:
-                        da = transcoder._encode(rich_act, context)
-                    else:
-                        da = transcoder.encode(rich_act, context)
-                    pack_into_act(
-                        data,
-                        act_off,
-                        da._location_id,
-                        da._text_string_id,
-                        da._wav_string_id,
-                        da._time,
-                        da._first_group,
-                        da._second_group,
-                        da._action_argument_type,
-                        da._action_id,
-                        da._quantifier_or_switch_or_order,
-                        da._flags,
-                        da._padding,
-                        da._mask_flag,
-                    )
-                act_off += num_bytes_per_act
-
-            # --- player execution ---
-            pe_base = act_base + acts_sz
-            key = trigger._players
-            pe_bytes = pe_cache.get(key)
-            if pe_bytes is None:
-                pe = pe_execution_cache.get(key)
-                if pe is None:
-                    player_flags = [0] * 27
-                    for player_id in trigger._players:
-                        player_flags[player_id._id] = 1
-                    pe = DecodedPlayerExecution(
-                        _execution_flags=0,
-                        _player_flags=player_flags,
-                        _current_action_index=0,
-                    )
-                    pe_execution_cache[key] = pe
-                buf = bytearray(pe_sz)
-                pack_into_pe(
-                    buf,
-                    0,
-                    pe._execution_flags,
-                    *pe._player_flags,
-                    pe._current_action_index,
-                )
-                pe_bytes = bytes(buf)
-                pe_cache[key] = pe_bytes
-            data[pe_base : pe_base + pe_sz] = pe_bytes
-
-            offset += trig_sz
-
-        return data
+    def get_secondary_cache_key(
+        self, rich_chk_section: RichTrigSection
+    ) -> Optional[tuple[Any, ...]]:
+        return self._optimizer.get_secondary_cache_key(rich_chk_section)
 
     def _encode_trigger(
         self, rich_trigger: RichTrigger, rich_chk_encode_context: RichChkEncodeContext
     ) -> DecodedTrigger:
         conditions = self._encode_conditions(
-            rich_trigger._conditions, rich_chk_encode_context
+            rich_trigger.conditions, rich_chk_encode_context
         )
-        actions = self._encode_actions(rich_trigger._actions, rich_chk_encode_context)
-        player_execution = self._encode_player_execution(rich_trigger._players)
+        actions = self._encode_actions(rich_trigger.actions, rich_chk_encode_context)
+        player_execution = self._encode_player_execution(rich_trigger.players)
         return DecodedTrigger(
             _conditions=conditions, _actions=actions, _player_execution=player_execution
         )
